feat: use ExprContext in SigCompleter (#10390)
